+++ airlinesystem/src/main/java/com/airline/airlinesystem/core/Flight.java
@@ -11,7 +11,7 @@ import jakarta.persistence.*;
 public class Flight {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id; // Primary key do not include in constructor
+    private int id; // Primary key do not include in constructor
 
     private String flightNo;
     private String destination;
@@ -220,11 +220,11 @@ public class Flight {
         return availableSeats;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
